+++ spec/models/user_spec.rb
@@ -220,6 +220,21 @@ describe User do
               2.times { user.rebrag! }
             end
           end
+
+          context 'with an unbragged activity (deleted from Strava)' do
+            before do
+              user.activities.bragged.each { |a| a.update_attributes!(unbragged_at: Time.now.utc) }
+            end
+
+            it 'does not include unbragged activities in latest_bragged_activity' do
+              expect(user.send(:latest_bragged_activity)).to be_nil
+            end
+
+            it 'does not rebrag the unbragged activity' do
+              expect(user).not_to receive(:rebrag_activity!)
+              user.rebrag!
+            end
+          end
         end
 
         context 'without a refresh token (until October 2019)', vcr: { cassette_name: 'strava/refresh_access_token' } do